seperate RealTimeWebsocketProcess class from ClientAsset class
